handleInput file renamed back to misc
