+++ client/src/utils/misc.js
@@ -3,8 +3,6 @@
 * Takes an object as update and setInputs function
 * */
 
-const handleInput = (update, setInputs) => {
+export const handleInput = (update, setInputs) => {
     return setInputs(prevState => ({ ...prevState, ...update }));
 }
-
-export default handleInput;
